questao 2 completa

+++ aula pratica09/questao02/Carro.java
@@ -3,7 +3,7 @@ public class Carro extends ProdutoDuravel
 	private String estilo;
 	private String portas;
 
-	public Carro(String nomeProduto, float preçoProduto, String marcaProduto,String descriçaoProduto,
+	public Carro(String nomeProduto, double preçoProduto, String marcaProduto,String descriçaoProduto,
 				 String dataDeFabricaçaoProduto, String mpp, String durabilidadeProduto, 
 				 String estiloCarro, String portasCarro)
 	{
