+++ aula pratica09/questao02/Chocolate.java
@@ -2,7 +2,7 @@ public class Chocolate extends ProdutoNaoDuravel
 {
 	private String sabor;
 
-	public Chocolate(String nomeProduto, float preçoProduto, String marcaProduto, String descriçaoProduto,
+	public Chocolate(String nomeProduto, double preçoProduto, String marcaProduto, String descriçaoProduto,
 				 	String dataDeFabricaçaoProduto, String data, String generoProduto, String saborChoco)
 	{
 		super(nomeProduto,preçoProduto,marcaProduto,descriçaoProduto,dataDeFabricaçaoProduto,
