+++ aula pratica09/questao02/Livro.java
@@ -3,7 +3,7 @@ public class Livro extends ProdutoDuravel
 	private String autor;
 	private int paginas;
 
-	public Livro(String nomeProduto, float preçoProduto, String marcaProduto,String descriçaoProduto,
+	public Livro(String nomeProduto, double preçoProduto, String marcaProduto,String descriçaoProduto,
 				 String dataDeFabricaçaoProduto, String mpp, String durabilidadeProduto, 
 				 String autorLivro, int paginasLivro)
 	{
