+++ aula pratica09/questao02/Pizza.java
@@ -3,7 +3,7 @@ public class Pizza extends ProdutoNaoDuravel
 	private String sabor;
 	private int fatias;
 
-	public Pizza(String nomeProduto, float preçoProduto, String marcaProduto, String descriçaoProduto,
+	public Pizza(String nomeProduto, double preçoProduto, String marcaProduto, String descriçaoProduto,
 				 String dataDeFabricaçaoProduto, String data, String generoProduto, String saborPizza, int fatiasPizza)
 	{
 		super(nomeProduto,preçoProduto,marcaProduto,descriçaoProduto,dataDeFabricaçaoProduto,
